+++ packages/fleather/lib/fleather.dart
@@ -15,7 +15,6 @@ export 'src/widgets/cursor.dart';
 export 'src/widgets/editor.dart';
 export 'src/widgets/editor_toolbar.dart';
 export 'src/widgets/field.dart';
-export 'src/widgets/history.dart' show HistoryStack;
 export 'src/widgets/link.dart' show LinkActionPickerDelegate, LinkMenuAction;
 export 'src/widgets/text_line.dart';
 export 'src/widgets/theme.dart';
